// src/state/types.ts
export interface TeamsState {
  isInitialized: boolean
  isLoading: boolean
  data: TeamState
}

export interface TeamOwner {
  name: string
  address: string
  avatar: NftImages
  vipBranch: boolean
}

export interface TeamState {
  name: string
  address: string
  users: number
  totalCommission: BigNumber
  avatar: NftImages
  points: BigNumber
  referralAddresses
  referralList: Array<TeamData>
  nftCommission: BigNumber
  owner: TeamOwner
  isVip: boolean
}


export interface TeamData {
  name: string
  users: number
  points: BigNumber
  totalFundSubTeam: BigNumber
  images: NftImages
}